init serviceList test
